Projects/scripts/requirements_linux.txt
botocore==1.25.0
boto3==1.22.0
pyserial
halo
cryptography
ecdsa
pyopenssl